// client/src/entities/interview/consts.ts
export enum STATUSES {
  ACTIVE = 'active',
  ARCHIVED = 'archived',
}

export enum RATING_TYPE {
  ALGORITHMS = 'algorithms',
  BASICKNOWLEDGE = 'basicKnowledge',
  CODEQUALITY = 'codeQuality',
  COMMUNICATION = 'communication',
  DECOMPOSE = 'decompose',
  DESIGN = 'design',
  REASONING = 'reasoning',
}